backend/requirements.txt
blinker==1.9.0
click==8.1.8
flask==3.1.1
flask-cors==6.0.0
flask-sqlalchemy==3.1.1
greenlet==3.2.2
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlalchemy==2.0.41
typing-extensions==4.14.0
werkzeug==3.1.3
zipp==3.22.0
